requirements.txt
torch>=1.10.0
torchvision
numpy
scikit-learn
trimesh>=3.9.0
Pillow
tqdm
wandb #Optional for logging
